+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sphinx-labs/plugins
 
+## 0.23.1
+
+### Patch Changes
+
+- 431d6ef0: Use broadcast file instead of events to get transaction receipts for deployment artifacts
+- b938c2d8: Update remappings and dependencies in quickstart guide
+- f2c5d280: Remove @eth-optimism/contracts-bedrock dependency due to a breaking change in a minor version update in their package
+- Updated dependencies [431d6ef0]
+- Updated dependencies [f2c5d280]
+  - @sphinx-labs/core@0.18.1
+  - @sphinx-labs/contracts@0.13.1
+
 ## 0.23.0
 
 ### Minor Changes